+++ addons/scene_manager/ignore_item.gd
@@ -3,18 +3,22 @@ extends HBoxContainer
 
 @onready var _root: Node = self
 
+# Finds and fills `_root` variable properly
 func _ready() -> void:
 	while true:
-		if _root.name == "Scene Manager" || _root.name == "root_container":
+		if _root.name == "Scene Manager" || _root.name == "menu":
 			break
 		_root = _root.get_parent()
 
+# Sets address of current ignore item
 func set_address(addr: String) -> void:
 	get_node("address").text = addr
 	name = addr
 
+# Returns address of current ignore item
 func get_address() -> String:
 	return get_node("address").text
 
+# Remove Button
 func _on_remove_button_up() -> void:
 	_root.emit_signal("delete_ignore_child", self)