+++ CHANGELOG.rst
@@ -4,9 +4,13 @@ CHANGELOG
 UNRELEASED
 ----------
 
+
+3.0.2 (2022-12-29)
+------------------
+
 * fix: change_scene function was changed to change_scene_to_file in conversion to godot4
-* docs: documentation added for every function in manager.gd script
 * fix: bugs on duplication check fixed
+* docs: documentation added for all functions
 
 3.0.1 (2022-12-28)
 ------------------